apps/icc-scraper/requirements.txt
annotated-types==0.7.0
certifi==2024.8.30
charset-normalizer==3.3.2
idna==3.8
Jinja2==3.1.4
MarkupSafe==2.1.5
numpy==2.1.1
pandas==2.2.2
pydantic==2.9.0
pydantic_core==2.23.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
requests==2.32.3
ruff==0.6.4
six==1.16.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
